+++ collections/src/test/kotlin/ktx/collections/arraysTest.kt
@@ -365,6 +365,38 @@ class ArraysTest {
   }
 
   @Test
+  fun `should remove elements from existing GdxArray`() {
+    val array = GdxArray.with(1, 2, 3, 4, 5)
+    array.removeAll { it > 10 }
+    assertEquals(GdxArray.with(1, 2, 3, 4, 5), array)
+
+    array.removeAll { it % 2 == 0 }
+    assertEquals(GdxArray.with(1, 3, 5), array)
+
+    array.removeAll { it is Number }
+    assertEquals(GdxArray<Int>(), array)
+
+    array.removeAll { it > 0 }
+    assertEquals(GdxArray<Int>(), array)
+  }
+
+  @Test
+  fun `should retain elements from existing GdxArray`() {
+    val array = GdxArray.with(1, 2, 3, 4, 5)
+    array.retainAll { it < 6 }
+    assertEquals(GdxArray.with(1, 2, 3, 4, 5), array)
+
+    array.retainAll { it % 2 == 1 }
+    assertEquals(GdxArray.with(1, 3, 5), array)
+
+    array.retainAll { it < 0 }
+    assertEquals(GdxArray<Int>(), array)
+
+    array.retainAll { it > 0 }
+    assertEquals(GdxArray<Int>(), array)
+  }
+
+  @Test
   fun `should map elements into a new GdxArray`() {
     val array = GdxArray.with(1, 2, 3)
 
@@ -372,7 +404,6 @@ class ArraysTest {
 
     assertEquals(GdxArray.with(2, 4, 6), result)
   }
-
   @Test
   fun `should filter elements into a new GdxArray`() {
     val array = GdxArray.with(1, 2, 3, 4, 5)